Add queries for autocomplete service
The rational behind having for two queries for autocomplete is that most APIs don't provide a complete address/place object for the autocomplete route, so N+1 requests would be necessary to get all addresses required for the suggestAddress query.

By making having a `externalId` attribute, the `suggestAddresses` query will just return a special id that will be followed up by a lookup query on the selected address. Basically, the flow for google would be:

```
*User starts typing*

                  suggestAddresses(searchTerm)
place-components -----------------------------> places-graphql
place-components <----------------------------- places-graphql
                  AddressSuggestion[] where
				  each element has an externalId

*User selects a suggestion*

                   getAddressByExternalId(id)
place-components ----------------------------> places-graphql
place-components <---------------------------- places-graphql
                            Address

And this returned address will be set on the Context
```